Move RateLimitScope to commands.ratelimit
