+++ src/main/java/com/freya02/botcommands/api/commands/ratelimit/RateLimitScope.java
@@ -1,6 +1,4 @@
-package com.freya02.botcommands.api.commands;
-
-import com.freya02.botcommands.api.commands.ratelimit.DefaultRateLimiter;
+package com.freya02.botcommands.api.commands.ratelimit;
 
 /**
  * Provides rate limit scopes for the {@link DefaultRateLimiter default rate limit implementation}.
