+++ src/main/kotlin/com/freya02/botcommands/api/commands/annotations/RateLimit.kt
@@ -1,7 +1,7 @@
 package com.freya02.botcommands.api.commands.annotations
 
-import com.freya02.botcommands.api.commands.RateLimitScope
 import com.freya02.botcommands.api.commands.ratelimit.RateLimitContainer
+import com.freya02.botcommands.api.commands.ratelimit.RateLimitScope
 import io.github.bucket4j.Bucket
 import io.github.bucket4j.ConsumptionProbe
 import java.time.temporal.ChronoUnit
