+++ src/main/kotlin/com/freya02/botcommands/api/commands/ratelimit/RateLimiter.kt
@@ -1,6 +1,5 @@
 package com.freya02.botcommands.api.commands.ratelimit
 
-import com.freya02.botcommands.api.commands.RateLimitScope
 import com.freya02.botcommands.api.commands.ratelimit.RateLimiter.Companion.defaultFactory
 import com.freya02.botcommands.api.commands.ratelimit.bucket.BucketAccessor
 import com.freya02.botcommands.api.commands.ratelimit.handler.RateLimitHandler
