+++ src/test/java/com/freya02/botcommands/test/commands/slash/SlashCooldown.java
@@ -1,10 +1,10 @@
 package com.freya02.botcommands.test.commands.slash;
 
-import com.freya02.botcommands.api.commands.RateLimitScope;
 import com.freya02.botcommands.api.commands.annotations.Cooldown;
 import com.freya02.botcommands.api.commands.application.ApplicationCommand;
 import com.freya02.botcommands.api.commands.application.slash.GuildSlashEvent;
 import com.freya02.botcommands.api.commands.application.slash.annotations.JDASlashCommand;
+import com.freya02.botcommands.api.commands.ratelimit.RateLimitScope;
 
 import java.time.temporal.ChronoUnit;
