+++ src/test/kotlin/com/freya02/botcommands/test_kt/commands/slash/SlashCooldown.kt
@@ -1,6 +1,5 @@
 package com.freya02.botcommands.test_kt.commands.slash
 
-import com.freya02.botcommands.api.commands.RateLimitScope
 import com.freya02.botcommands.api.commands.annotations.Command
 import com.freya02.botcommands.api.commands.annotations.Cooldown
 import com.freya02.botcommands.api.commands.application.ApplicationCommand
@@ -9,6 +8,7 @@ import com.freya02.botcommands.api.commands.application.annotations.AppDeclarati
 import com.freya02.botcommands.api.commands.application.slash.GuildSlashEvent
 import com.freya02.botcommands.api.commands.application.slash.annotations.JDASlashCommand
 import com.freya02.botcommands.api.commands.builder.cooldown
+import com.freya02.botcommands.api.commands.ratelimit.RateLimitScope
 import dev.minn.jda.ktx.messages.reply_
 import java.time.temporal.ChronoUnit
 import kotlin.time.Duration.Companion.seconds
