+++ src/test/kotlin/com/freya02/botcommands/test_kt/commands/slash/SlashMyJavaCommand.java
@@ -2,7 +2,6 @@ package com.freya02.botcommands.test_kt.commands.slash;
 
 import com.freya02.botcommands.api.BContext;
 import com.freya02.botcommands.api.commands.CommandPath;
-import com.freya02.botcommands.api.commands.RateLimitScope;
 import com.freya02.botcommands.api.commands.annotations.*;
 import com.freya02.botcommands.api.commands.application.ApplicationCommand;
 import com.freya02.botcommands.api.commands.application.slash.ApplicationGeneratedValueSupplier;
@@ -13,6 +12,7 @@ import com.freya02.botcommands.api.commands.application.slash.annotations.SlashO
 import com.freya02.botcommands.api.commands.application.slash.autocomplete.AutocompleteCacheMode;
 import com.freya02.botcommands.api.commands.application.slash.autocomplete.annotations.AutocompleteHandler;
 import com.freya02.botcommands.api.commands.application.slash.autocomplete.annotations.CacheAutocomplete;
+import com.freya02.botcommands.api.commands.ratelimit.RateLimitScope;
 import com.freya02.botcommands.api.parameters.ParameterType;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.User;
